feat: add spec discovery, expand MCP API, and enhance execution pipeline
Add spec discovery tab and spec execution libraries for the frontend.
Expand MCP API with new endpoints, add spec and error_resolved step
handlers, improve orchestrator knowledge building and verification,
enhance summary generator, add god class analyzers for Rust and
TypeScript, refactor Chrome extension with offscreen support, and
update unified workflow executor with improved loop control and phases.
Remove deprecated LivePageGenerator component.

+++ src-tauri/src/error_monitor/curator.rs
@@ -279,7 +279,7 @@ impl DebugContextCurator {
         }
 
         // Occurrence count matters
-        score += (error.occurrence_count.min(20) * 3) as u32;
+        score += error.occurrence_count.min(20) * 3;
 
         // Newer errors are more relevant
         if error.status == ErrorStatus::New {
