+++ src-tauri/src/step_executor/handlers/awas_execute.rs
@@ -29,7 +29,7 @@ impl StepHandler for AwasExecuteHandler {
         context: &HandlerContext,
     ) -> StepHandlerResult {
         let step_name = step.name.as_deref().unwrap_or("AWAS Execute");
-        let timeout = step.timeout_seconds.unwrap_or(60) as u64;
+        let timeout = step.timeout_seconds.unwrap_or(60);
 
         // Get URL from step config
         let url = match &step.awas_url {
